Move the button promise logic to a hook
